Refactor: move clip_preprocess to comfy.clip_model (#11586)
